Reliably set the quickfix title
Right now w:quickfix_title is reset after closing the quickfix window.
As I see it, setting it doesn't associate it with the current quickfix list.
I don't know if such behavior should be classified as a bug in Vim.
For the time being, there is an easy fix for this.