Add a [fname] to the doautocmd command
The pattern in QuickFixCmdPre and QuickFixCmdPost autocommands is
matched against the command being run. Since vim-qf uses l* and [^l]* to
distinguish between commands that perform on the quickfix and location
list, we need to add a [fname] to our "doautocmd {event} [fname]" call
as well. Vim will otherwise use the current file name which will lead to
an error for file names starting with an 'l', since vim-qlist always
populates the quickfix list. Using 'cwindow' as [fname] makes it more
obvious what we're trying to trigger.